tensor more, move validate_broadcast_all_sources
